First cut of dazzler overlaps, blasr alignments
Using the first and last overlap from a linear set between a/b in one
direction, parse the substrings and pass them to SimpleAligner, which
calls blasrs sparse-dynamic-programming alignment.  This gives us the
speed of daligner overlapping with the forgiveness of blasr alignment
through problematic regions and ultimately save the integrity of the
seed read.

More testing needed.